+++ src/aks-preview/azext_aks_preview/tests/latest/test_managed_cluster_decorator.py
@@ -584,7 +584,7 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
         )
         load_balancer_profile_2 = self.models.load_balancer_models.ManagedClusterLoadBalancerProfile(
             managed_outbound_i_ps=self.models.load_balancer_models.ManagedClusterLoadBalancerProfileManagedOutboundIPs(
-                count=10, count_i_pv6=20
+                count=10, count_ipv6=20
             ),
             outbound_i_ps=self.models.load_balancer_models.ManagedClusterLoadBalancerProfileOutboundIPs(
                 public_i_ps=[
@@ -625,7 +625,7 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
         self.assertEqual(ctx_1.get_load_balancer_managed_outbound_ipv6_count(), None)
         load_balancer_profile = self.models.load_balancer_models.ManagedClusterLoadBalancerProfile(
             managed_outbound_i_ps=self.models.load_balancer_models.ManagedClusterLoadBalancerProfileManagedOutboundIPs(
-                count_i_pv6=10
+                count_ipv6=10
             )
         )
         network_profile = self.models.ContainerServiceNetworkProfile(
@@ -663,7 +663,7 @@ class AKSPreviewManagedClusterContextTestCase(unittest.TestCase):
         )
         load_balancer_profile_3 = self.models.load_balancer_models.ManagedClusterLoadBalancerProfile(
             managed_outbound_i_ps=self.models.load_balancer_models.ManagedClusterLoadBalancerProfileManagedOutboundIPs(
-                count=10, count_i_pv6=20
+                count=10, count_ipv6=20
             ),
             outbound_i_ps=self.models.load_balancer_models.ManagedClusterLoadBalancerProfileOutboundIPs(
                 public_i_ps=[
@@ -10149,7 +10149,7 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
             network_profile=self.models.ContainerServiceNetworkProfile(
                 load_balancer_profile=self.models.load_balancer_models.ManagedClusterLoadBalancerProfile(
                     managed_outbound_i_ps=self.models.load_balancer_models.ManagedClusterLoadBalancerProfileManagedOutboundIPs(
-                        count=10, count_i_pv6=20
+                        count=10, count_ipv6=20
                     ),
                 )
             ),
@@ -10163,7 +10163,7 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
                 self.models.ContainerServiceNetworkProfile(
                     load_balancer_profile=self.models.load_balancer_models.ManagedClusterLoadBalancerProfile(
                         managed_outbound_i_ps=self.models.load_balancer_models.ManagedClusterLoadBalancerProfileManagedOutboundIPs(
-                            count=5, count_i_pv6=20
+                            count=5, count_ipv6=20
                         ),
                     )
                 )
@@ -10189,7 +10189,7 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
             network_profile=self.models.ContainerServiceNetworkProfile(
                 load_balancer_profile=self.models.load_balancer_models.ManagedClusterLoadBalancerProfile(
                     managed_outbound_i_ps=self.models.load_balancer_models.ManagedClusterLoadBalancerProfileManagedOutboundIPs(
-                        count=10, count_i_pv6=20
+                        count=10, count_ipv6=20
                     ),
                 )
             ),
@@ -10203,7 +10203,7 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
                 self.models.ContainerServiceNetworkProfile(
                     load_balancer_profile=self.models.load_balancer_models.ManagedClusterLoadBalancerProfile(
                         managed_outbound_i_ps=self.models.load_balancer_models.ManagedClusterLoadBalancerProfileManagedOutboundIPs(
-                            count=10, count_i_pv6=5
+                            count=10, count_ipv6=5
                         ),
                     )
                 )
@@ -10229,7 +10229,7 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
             network_profile=self.models.ContainerServiceNetworkProfile(
                 load_balancer_profile=self.models.load_balancer_models.ManagedClusterLoadBalancerProfile(
                     managed_outbound_i_ps=self.models.load_balancer_models.ManagedClusterLoadBalancerProfileManagedOutboundIPs(
-                        count=10, count_i_pv6=20
+                        count=10, count_ipv6=20
                     ),
                 )
             ),
@@ -10243,7 +10243,7 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
                 self.models.ContainerServiceNetworkProfile(
                     load_balancer_profile=self.models.load_balancer_models.ManagedClusterLoadBalancerProfile(
                         managed_outbound_i_ps=self.models.load_balancer_models.ManagedClusterLoadBalancerProfileManagedOutboundIPs(
-                            count=25, count_i_pv6=5
+                            count=25, count_ipv6=5
                         ),
                     )
                 )
@@ -10268,7 +10268,7 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
             network_profile=self.models.ContainerServiceNetworkProfile(
                 load_balancer_profile=self.models.load_balancer_models.ManagedClusterLoadBalancerProfile(
                     managed_outbound_i_ps=self.models.load_balancer_models.ManagedClusterLoadBalancerProfileManagedOutboundIPs(
-                        count=3, count_i_pv6=2
+                        count=3, count_ipv6=2
                     )
                 )
             ),
@@ -10330,7 +10330,7 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
 
         ground_truth_load_balancer_profile_8 = self.models.load_balancer_models.ManagedClusterLoadBalancerProfile(
             managed_outbound_i_ps=self.models.load_balancer_models.ManagedClusterLoadBalancerProfileManagedOutboundIPs(
-                count=10, count_i_pv6=5
+                count=10, count_ipv6=5
             ),
         )
         ground_truth_network_profile_8 = self.models.ContainerServiceNetworkProfile(
@@ -16841,10 +16841,8 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
         ground_truth_mc_1 = self.models.ManagedCluster(
             location="test_location",
             scheduler_profile=self.models.SchedulerProfile(
-                scheduler_instance_profiles=self.models.SchedulerProfileSchedulerInstanceProfiles(
-                    upstream=self.models.SchedulerInstanceProfile(
-                        scheduler_config_mode=self.models.SchedulerConfigMode.MANAGED_BY_CRD
-                    )
+                upstream=self.models.SchedulerInstanceProfile(
+                    scheduler_config_mode=self.models.SchedulerConfigMode.MANAGED_BY_CRD
                 )
             ),
         )
@@ -16862,10 +16860,8 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
         mc_2 = self.models.ManagedCluster(
             location="test_location",
             scheduler_profile=self.models.SchedulerProfile(
-                scheduler_instance_profiles=self.models.SchedulerProfileSchedulerInstanceProfiles(
-                    upstream=self.models.SchedulerInstanceProfile(
-                        scheduler_config_mode=self.models.SchedulerConfigMode.DEFAULT
-                    )
+                upstream=self.models.SchedulerInstanceProfile(
+                    scheduler_config_mode=self.models.SchedulerConfigMode.DEFAULT
                 )
             ),
         )
@@ -16874,10 +16870,8 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
         ground_truth_mc_2 = self.models.ManagedCluster(
             location="test_location",
             scheduler_profile=self.models.SchedulerProfile(
-                scheduler_instance_profiles=self.models.SchedulerProfileSchedulerInstanceProfiles(
-                    upstream=self.models.SchedulerInstanceProfile(
-                        scheduler_config_mode=self.models.SchedulerConfigMode.MANAGED_BY_CRD
-                    )
+                upstream=self.models.SchedulerInstanceProfile(
+                    scheduler_config_mode=self.models.SchedulerConfigMode.MANAGED_BY_CRD
                 )
             ),
         )
@@ -16912,10 +16906,8 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
         ground_truth_mc_1 = self.models.ManagedCluster(
             location="test_location",
             scheduler_profile=self.models.SchedulerProfile(
-                scheduler_instance_profiles=self.models.SchedulerProfileSchedulerInstanceProfiles(
-                    upstream=self.models.SchedulerInstanceProfile(
-                        scheduler_config_mode=self.models.SchedulerConfigMode.MANAGED_BY_CRD
-                    )
+                upstream=self.models.SchedulerInstanceProfile(
+                    scheduler_config_mode=self.models.SchedulerConfigMode.MANAGED_BY_CRD
                 )
             ),
         )
@@ -16936,10 +16928,8 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
         ground_truth_mc_2 = self.models.ManagedCluster(
             location="test_location",
             scheduler_profile=self.models.SchedulerProfile(
-                scheduler_instance_profiles=self.models.SchedulerProfileSchedulerInstanceProfiles(
-                    upstream=self.models.SchedulerInstanceProfile(
-                        scheduler_config_mode=self.models.SchedulerConfigMode.DEFAULT
-                    )
+                upstream=self.models.SchedulerInstanceProfile(
+                    scheduler_config_mode=self.models.SchedulerConfigMode.DEFAULT
                 )
             ),
         )
@@ -16972,10 +16962,8 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
         mc_4 = self.models.ManagedCluster(
             location="test_location",
             scheduler_profile=self.models.SchedulerProfile(
-                scheduler_instance_profiles=self.models.SchedulerProfileSchedulerInstanceProfiles(
-                    upstream=self.models.SchedulerInstanceProfile(
-                        scheduler_config_mode=self.models.SchedulerConfigMode.DEFAULT
-                    )
+                upstream=self.models.SchedulerInstanceProfile(
+                    scheduler_config_mode=self.models.SchedulerConfigMode.DEFAULT
                 )
             ),
         )
@@ -16984,10 +16972,8 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
         ground_truth_mc_4 = self.models.ManagedCluster(
             location="test_location",
             scheduler_profile=self.models.SchedulerProfile(
-                scheduler_instance_profiles=self.models.SchedulerProfileSchedulerInstanceProfiles(
-                    upstream=self.models.SchedulerInstanceProfile(
-                        scheduler_config_mode=self.models.SchedulerConfigMode.MANAGED_BY_CRD
-                    )
+                upstream=self.models.SchedulerInstanceProfile(
+                    scheduler_config_mode=self.models.SchedulerConfigMode.MANAGED_BY_CRD
                 )
             ),
         )
@@ -17005,10 +16991,8 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
         mc_5 = self.models.ManagedCluster(
             location="test_location",
             scheduler_profile=self.models.SchedulerProfile(
-                scheduler_instance_profiles=self.models.SchedulerProfileSchedulerInstanceProfiles(
-                    upstream=self.models.SchedulerInstanceProfile(
-                        scheduler_config_mode=self.models.SchedulerConfigMode.MANAGED_BY_CRD
-                    )
+                upstream=self.models.SchedulerInstanceProfile(
+                    scheduler_config_mode=self.models.SchedulerConfigMode.MANAGED_BY_CRD
                 )
             ),
         )
@@ -17017,10 +17001,8 @@ class AKSPreviewManagedClusterUpdateDecoratorTestCase(unittest.TestCase):
         ground_truth_mc_5 = self.models.ManagedCluster(
             location="test_location",
             scheduler_profile=self.models.SchedulerProfile(
-                scheduler_instance_profiles=self.models.SchedulerProfileSchedulerInstanceProfiles(
-                    upstream=self.models.SchedulerInstanceProfile(
-                        scheduler_config_mode=self.models.SchedulerConfigMode.DEFAULT
-                    )
+                upstream=self.models.SchedulerInstanceProfile(
+                    scheduler_config_mode=self.models.SchedulerConfigMode.DEFAULT
                 )
             ),
         )