+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/cluster/_inspect.py
@@ -22,9 +22,9 @@ class Inspect(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2026-05-01-preview",
+        "version": "2026-07-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/clusters/{}/inspect", "2026-05-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/clusters/{}/inspect", "2026-07-01"],
         ]
     }
 
@@ -175,7 +175,7 @@ class Inspect(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2026-05-01-preview",
+                    "api-version", "2026-07-01",
                     required=True,
                 ),
             }