+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/clustermanager/identity/__init__.py
@@ -12,4 +12,3 @@ from .__cmd_group import *
 from ._assign import *
 from ._remove import *
 from ._show import *
-from ._wait import *